requirements_txt.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
matplotlib>=3.5.0